reorg; tests in tests dir.
