+++ rust/src/lib/mod.rs
@@ -1,5 +1,6 @@
 pub mod ast;
 pub mod cbpv;
+pub mod check;
 pub mod format;
 #[allow(clippy::all)]
 pub mod parser;
